refactor(whispering): rename DbService to BlobStore
DbService is now purely an audio blob store—no transformations, no
runs, no metadata. The name finally matches the scope:

- DbService → BlobStore, DbError → BlobError
- DbServiceLive → BlobStoreLive
- services.db → services.blobs
- services/db/ → services/blob-store/
- createDbServiceDesktop → createBlobStoreDesktop
- createDbServiceWeb → createBlobStoreWeb
- createFileSystemDbService → createFileSystemBlobStore
