+++ apps/whispering/src/lib/query/audio.ts
@@ -10,11 +10,11 @@ export const audio = {
 	/**
 	 * Get audio playback URL for a recording by ID.
 	 * Audio blobs are too large for Yjs CRDTs, so they're still served
-	 * from Dexie (web) / filesystem (desktop) via DbService.
+	 * from Dexie (web) / filesystem (desktop) via BlobStore.
 	 */
 	getPlaybackUrl: (id: Accessor<string>) =>
 		defineQuery({
 			queryKey: audioKeys.playbackUrl(id()),
-			queryFn: () => services.db.audio.ensurePlaybackUrl(id()),
+			queryFn: () => services.blobs.audio.ensurePlaybackUrl(id()),
 		}),
 };
