+++ apps/whispering/src/lib/services/blob-store/file-system.ts
@@ -1,77 +1,184 @@
-import { readDir, readFile, remove, writeFile } from '@tauri-apps/plugin-fs';
-import type { BlobStore } from './types.js';
+import { convertFileSrc, invoke } from '@tauri-apps/api/core';
+import {
+	exists,
+	mkdir,
+	readDir,
+	writeFile as tauriWriteFile,
+} from '@tauri-apps/plugin-fs';
+import mime from 'mime';
+import { tryAsync } from 'wellcrafted/result';
+import { PATHS } from '$lib/constants/paths';
+import { FsServiceLive } from '$lib/services/desktop/fs';
+import type { BlobStore } from './types';
+import { BlobError } from './types';
 
-const EXTENSION_TO_MIME: Record<string, string> = {
-	webm: 'audio/webm',
-	mp3: 'audio/mpeg',
-	wav: 'audio/wav',
-	ogg: 'audio/ogg',
-	mp4: 'audio/mp4',
-	m4a: 'audio/mp4',
-};
-
-const MIME_TO_EXTENSION: Record<string, string> = {
-	'audio/webm': 'webm',
-	'audio/mpeg': 'mp3',
-	'audio/wav': 'wav',
-	'audio/ogg': 'ogg',
-	'audio/mp4': 'mp4',
-};
-
-function mimeFromExtension(ext: string): string {
-	return EXTENSION_TO_MIME[ext] ?? 'application/octet-stream';
-}
-
-function extensionFromMime(mimeType: string): string {
-	return MIME_TO_EXTENSION[mimeType] ?? 'bin';
+/**
+ * Reads all markdown files from a directory using the Rust command.
+ * This is a single FFI call that reads all .md files natively in Rust,
+ * avoiding thousands of individual async calls for path joining and file reading.
+ *
+ * @param directoryPath - Absolute path to the directory containing .md files
+ * @returns Array of markdown file contents as strings
+ */
+export async function readMarkdownFiles(
+	directoryPath: string,
+): Promise<string[]> {
+	return invoke('read_markdown_files', { directoryPath });
 }
 
 /**
- * Find all files in `basePath` whose name starts with `{id}.`.
+ * Deletes files inside a directory by filename.
+ * Validates that filenames are single path components (no traversal).
+ *
+ * @param directory - Absolute path to the directory containing the files
+ * @param filenames - Array of leaf filenames to delete
+ * @returns Number of files successfully deleted
  */
-async function findMatchingFiles(
-	basePath: string,
-	id: string,
-): Promise<string[]> {
-	const prefix = `${id}.`;
-	const entries = await readDir(basePath);
-	return entries
-		.filter((entry) => entry.name.startsWith(prefix))
-		.map((entry) => entry.name);
+async function deleteFilesInDirectory(
+	directory: string,
+	filenames: string[],
+): Promise<number> {
+	return invoke('delete_files_in_directory', { directory, filenames });
 }
 
-export function createFileSystemBlobStore(basePath: string): BlobStore {
+/**
+ * File system-based blob store implementation for desktop.
+ * Stores data as markdown files with YAML front matter.
+ *
+ * Directory structure:
+ * - recordings/
+ *   - {id}.{ext} (audio file: .wav, .opus, .mp3, etc.)
+ *   - {id}.md (metadata materialized by workspace, NOT written by this service)
+ * - transformations/
+ *   - {id}.md (transformation configuration)
+ * - transformation-runs/
+ *   - {id}.md (execution history)
+ */
+export function createFileSystemBlobStore(): BlobStore {
 	return {
-		async get(id) {
-			const matches = await findMatchingFiles(basePath, id);
-			if (matches.length === 0) return null;
+		audio: {
+			async save(recordingId, audio) {
+				return tryAsync({
+					try: async () => {
+						const recordingsPath = await PATHS.DB.RECORDINGS();
+						await mkdir(recordingsPath, { recursive: true });
 
-			const filename = matches[0]!;
-			const ext = filename.split('.').pop() ?? '';
-			const mimeType = mimeFromExtension(ext);
+						const extension = mime.getExtension(audio.type) ?? 'bin';
+						const audioPath = await PATHS.DB.RECORDING_AUDIO(
+							recordingId,
+							extension,
+						);
+						const arrayBuffer = await audio.arrayBuffer();
+						await tauriWriteFile(audioPath, new Uint8Array(arrayBuffer));
+					},
+					catch: (error) => BlobError.MutationFailed({ cause: error }),
+				});
+			},
 
-			const bytes = await readFile(`${basePath}/${filename}`);
-			const blob = new Blob([bytes], { type: mimeType });
+			async delete(idOrIds) {
+				const ids = Array.isArray(idOrIds) ? idOrIds : [idOrIds];
+				return tryAsync({
+					try: async () => {
+						const recordingsPath = await PATHS.DB.RECORDINGS();
+						const idsToDelete = new Set(ids);
+						const allFiles = await readDir(recordingsPath);
+						const filenames = allFiles
+							.filter((file) => {
+								const id = file.name.split('.')[0] ?? '';
+								return idsToDelete.has(id);
+							})
+							.map((file) => file.name);
+						await deleteFilesInDirectory(recordingsPath, filenames);
+					},
+					catch: (error) => BlobError.MutationFailed({ cause: error }),
+				});
+			},
 
-			return { blob, mimeType };
-		},
+			async getBlob(recordingId: string) {
+				return tryAsync({
+					try: async () => {
+						const recordingsPath = await PATHS.DB.RECORDINGS();
+						const audioFilename = await findAudioFile(
+							recordingsPath,
+							recordingId,
+						);
 
-		async put(id, blob, mimeType) {
-			const ext = extensionFromMime(mimeType);
-			const arrayBuffer = await blob.arrayBuffer();
-			await writeFile(`${basePath}/${id}.${ext}`, new Uint8Array(arrayBuffer));
-		},
+						if (!audioFilename) {
+							throw new Error(
+								`Audio file not found for recording ${recordingId}`,
+							);
+						}
 
-		async delete(id) {
-			const matches = await findMatchingFiles(basePath, id);
-			await Promise.all(
-				matches.map((filename) => remove(`${basePath}/${filename}`)),
-			);
-		},
+						const audioPath = await PATHS.DB.RECORDING_FILE(audioFilename);
 
-		async has(id) {
-			const matches = await findMatchingFiles(basePath, id);
-			return matches.length > 0;
+						// Use existing fsService.pathToBlob utility
+						const { data: blob, error } =
+							await FsServiceLive.pathToBlob(audioPath);
+						if (error) throw error;
+
+						return blob;
+					},
+					catch: (error) => BlobError.QueryFailed({ cause: error }),
+				});
+			},
+
+			async ensurePlaybackUrl(recordingId: string) {
+				return tryAsync({
+					try: async () => {
+						const recordingsPath = await PATHS.DB.RECORDINGS();
+						const audioFilename = await findAudioFile(
+							recordingsPath,
+							recordingId,
+						);
+
+						if (!audioFilename) {
+							throw new Error(
+								`Audio file not found for recording ${recordingId}`,
+							);
+						}
+
+						const audioPath = await PATHS.DB.RECORDING_FILE(audioFilename);
+						const assetUrl = convertFileSrc(audioPath);
+
+						// Return the URL as-is from convertFileSrc()
+						// The Tauri backend handles URL decoding automatically
+						return assetUrl;
+					},
+					catch: (error) => BlobError.QueryFailed({ cause: error }),
+				});
+			},
+
+			revokeUrl(_recordingId: string) {
+				// No-op on desktop, URLs are asset:// protocol managed by Tauri
+			},
+
+			async clear() {
+				return tryAsync({
+					try: async () => {
+						const recordingsPath = await PATHS.DB.RECORDINGS();
+						const dirExists = await exists(recordingsPath);
+						if (!dirExists) return undefined;
+
+						const allFiles = await readDir(recordingsPath);
+						const filenames = allFiles.map((file) => file.name);
+						await deleteFilesInDirectory(recordingsPath, filenames);
+					},
+					catch: (error) => BlobError.MutationFailed({ cause: error }),
+				});
+			},
 		},
 	};
+}
+
+/**
+ * Helper function to find audio file by ID.
+ * Reads directory once and finds the matching file by ID prefix.
+ * This is much faster than checking every possible extension.
+ */
+async function findAudioFile(dir: string, id: string): Promise<string | null> {
+	const files = await readDir(dir);
+	const audioFile = files.find(
+		(f) => f.name.startsWith(`${id}.`) && !f.name.endsWith('.md'),
+	);
+	return audioFile?.name ?? null;
 }
