+++ apps/whispering/src/lib/services/blob-store/index.ts
@@ -1,3 +1,10 @@
-export { createFileSystemBlobStore } from './file-system.js';
-export { createIndexedDbBlobStore } from './indexed-db.js';
-export type { BlobStore } from './types.js';
+import { DownloadServiceLive } from '$lib/services/download';
+import { createBlobStoreDesktop } from './desktop';
+import { createBlobStoreWeb } from './web';
+
+export type { BlobStore } from './types';
+export { BlobError } from './types';
+
+export const BlobStoreLive = window.__TAURI_INTERNALS__
+	? createBlobStoreDesktop({ DownloadService: DownloadServiceLive })
+	: createBlobStoreWeb({ DownloadService: DownloadServiceLive });
