+++ apps/whispering/src/routes/(app)/(config)/recordings/+page.svelte
@@ -169,7 +169,7 @@
 							description: 'Are you sure you want to delete this recording?',
 							confirm: { text: 'Delete', variant: 'destructive' },
 							onConfirm: () => {
-								services.db.audio.revokeUrl(row.original.id);
+								services.blobs.audio.revokeUrl(row.original.id);
 								recordings.delete(row.original.id);
 								rpc.notify.success({
 									title: 'Deleted recording!',
