+++ apps/whispering/src/routes/(app)/(config)/recordings/RenderAudioUrl.svelte
@@ -13,7 +13,7 @@
 
 	onDestroy(() => {
 		// Clean up audio URL when component unmounts to prevent memory leaks
-		services.db.audio.revokeUrl(id);
+		services.blobs.audio.revokeUrl(id);
 	});
 </script>
